Script Engine Update Phase 1
Dropping array limits, test phase.
Also Implements a under-development (u)int64 key support to the dbmaps.
Made Possible Thanks to Haruna!

+++ src/map/mapreg.h
@@ -6,7 +6,7 @@
 #define _MAPREG_H_
 
 struct mapreg_save {
-	int uid;
+	int64 uid;
 	union {
 		int i;
 		char *str;
@@ -19,9 +19,9 @@ void mapreg_final(void);
 void mapreg_init(void);
 bool mapreg_config_read(const char* w1, const char* w2);
 
-int mapreg_readreg(int uid);
-char* mapreg_readregstr(int uid);
-bool mapreg_setreg(int uid, int val);
-bool mapreg_setregstr(int uid, const char* str);
+int mapreg_readreg(int64 uid);
+char* mapreg_readregstr(int64 uid);
+bool mapreg_setreg(int64 uid, int val);
+bool mapreg_setregstr(int64 uid, const char* str);
 
 #endif /* _MAPREG_H_ */
